Fix types mismatch

+++ mobility/motives/leisure.py
@@ -65,7 +65,7 @@ class LeisureMotive(Motive):
                 opportunities[["transport_zone_id", "n_opp"]]
                 .rename({"transport_zone_id": "to"}, axis=1)
             )
-            opportunities["to"] = opportunities["to"].astype("Int64")
+            opportunities["to"] = opportunities["to"].astype("Int32")
             
             if os.environ.get("MOBILITY_DEBUG") == "1":
                 self.plot_opportunities_map(
